+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Bumped runtimeVersion to `2022.06.22`.
+ * Upgraded preview Flutter analysis SDK to `3.1.0-9.0.pre`.
 
 ## `20220620t083300-all`
  * Bumped runtimeVersion to `2022.06.13`.